+++ changelog.md
@@ -1,5 +1,9 @@
 
-# VeitzQueryTool
+# VeitzQueryTool - Changelog
+
+## CHANGELOG v2.2:
+- add trading tools - get real time BTC-value ... onetrading / binance
+- add trading tools - real time graph ... onetrading / binance
 
 ## CHANGELOG v2.0:
 - initial Release 2.0